+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stoker-platform/utils
 
+## 0.5.86
+
+### Patch Changes
+
+- chore: update dependencies
+- Updated dependencies
+    - @stoker-platform/types@0.5.69
+
 ## 0.5.85
 
 ### Patch Changes